% Generated by roxygen2 (4.0.1): do not edit by hand
\name{twilio_call}
\alias{twilio_call}
\title{Call someone with a TwiML Voice}
\usage{
twilio_call(accountSID, auth_token, xmlURL, from, to)
}
\arguments{
\item{accountSID}{Your account ID}

\item{auth_token}{Your secret token}

\item{xmlURL}{The URL where the TwiML file is located.}

\item{from}{The phonenumber that calls.  Check Your Twilio account for the number.
Make sure to include the international number.}

\item{to}{The phone number to call.  Make sure to include the international number.}
}
\description{
You can use this to call someone and use a pre-uploaded twilio XML (TwiML)
file with what is to be said.
}

