% Generated by roxygen2 (4.0.1): do not edit by hand
\name{twilio_stateless_call}
\alias{twilio_stateless_call}
\title{Send an outbound call without writing the file to disk}
\usage{
twilio_stateless_call(accountSID, auth_token, message = NULL, TwXML = NULL,
  From, To)
}
\arguments{
\item{accountSID}{Your account ID}

\item{auth_token}{Your secret token}

\item{message}{the message to transcribe to speech. By default this will be spoken by a female
in british english and repeated 3 times. \emph{Either message or Twxml must be null!
you cannot send a call with both parameters set}}

\item{TwXML}{Any properly formated XML acceptable by Twillio}

\item{from}{The phonenumber that calls. Check Your Twilio account for the number.
Make sure to include the international number.}

\item{to}{The phone number to call. Make sure to include the international number.}
}
\description{
this function sends an outbound call using a twillio twimlet,
allowing stateless application calling functionality
}

