% Generated by roxygen2 (4.0.1): do not edit by hand
\name{twilio_stateless_SMS}
\alias{twilio_stateless_SMS}
\title{Send a SMS text message from Twillio}
\usage{
twilio_stateless_SMS(accountSID, auth_token, message = NULL, From, To)
}
\arguments{
\item{accountSID}{Your account ID}

\item{auth_token}{Your secret token}

\item{message}{the message to send in  a text.}

\item{From}{The phonenumber that calls. Check Your Twilio account for the number.
Make sure to include the international number.}

\item{To}{The phone number to call. \emph{You can include multple numbers as a quoted charecter
vector to sent the same message to multiple phones at the same time}.
Make sure to include the international number.}
}
\description{
This function can send a text message to one or multiple numbers
}

